Modernize Psyshock
Documentation updates, in parameters, and Burst compatible layer building are just a few of the improvements in this massive review-and-fix. Others include bugfixes to PhysicsBufferFromEntity, ScaleCollider, and MPR internal checks.

An interesting improvement was the rearrangement of the Collider struct such that there is no longer special casing for types with blob assets.

Overall, Psyshock is in a much better place API-wise. The next thing will be to stress test some of underlying algorithms in a new environment. Time for Open Project 2!