Unity 2022 Conversion Checkpoint
So I think I have LSSS working with conversion. But I can't know for sure because rendering no longer works with conversion. But I can at least hear the battle, which is a good sign for Myri.

Kinemation is temporarily removed and not updated because it needs a thorough rewrite to conform to the new rendering stack.

Next up is migrating things to bakers in Latios Framework 0.4 style.