More 1.0 exp cleanup
BuildCollisionLayerInternal should now work correctly with enableable components in the passed in query. However, performance may not be great in those cases due to having to calculate the number of items in the query in advance.

Switching Myri's singular system over was one of the easiest conversions to ISystem to date. It just worked.