Kinemation Compiles for 1.0 experimental
I have no idea if any of this code actually works (Kinemation is not enabled in LSSS yet). But it at least compiles now, and all the modifications I believe needed to be made are made.

Also, I now am forcing Psyshock to use the Baker Bootstrap Installer, and LSSS still runs. So that functionality is working as intended. I also fixed the ComponentTypeHandle issue with BuildCollisionLayer. I forgot I made the handle caching in 0.50 and that didn't have EntityTypeHandle.