Add ParentChangeSystem
None of this is tested yet, so there's probably some bugs. But I am very happy with the design and what I believe this new system should be capable. It handles a lot more edge cases to fix the archetype of entities, all while parallelizing more of the workload.

I also fixed SafeEntity to respect Entity Index 0. I don't think that should ever be an issue in practice as usually WorldTime or WorldBlackboardEntity claim that index, but the fix was simple enough.

A future optimization could involve using EntityQueryMask for looking up whether an entity has a component. I'm not sure if that will actually be faster.

Next up is the TransformHierarchyUpdateSystem. I'm going to start by implementing ImprovedTransforms algorithm. I'll do Extreme Transforms once I have this whole Latios Transforms thing in a working state.