Latios Transforms for Psyshock queries
This is the first phase of the Psyshock migration to Latios Transforms.

The collider types which do not have a perfect stretching algorithm now have enums to specify how they stretch.

And now there are overloads for static methods which work with TransformQvvs types. I did not extend these to all the precise collider variants. I'm not sure if that is even something I want to do or if it is better to reduce the API to only take Collider types and always do dispatching.

I think I also fixed a bug with raycasting compound colliders where the results were in the wrong coordinate space.

Next up is to collapse everything to use only LocalToWorld. Then switch CollisionLayer to store TransformQvvs.