Psyshock Query Refactor
This was a big undertaking. Now all the queries are classified by the collider types involved rather than what they do. This makes it easier to optimize algorithms for specific collider types, as all optimizations and variations for a type or pair of types can exist in a dedicated static class for that type.

Another benefit of this approach is that is should be easier to incrementally develop new collider types, as it is much easier to see which internal method signatures are supported for a given type.

Next up is XML docs, and then I need to revisit PhysicsDebug.