Latios Transforms are working
LSSS is back with a whole new set of transforms! Basic scenes, stress tests, everything is operating as I would expect!

As for performance, it is still in regression territory in the stress-test scenes. But this is because there's a little less scheduling dynamics available for spawner transforms between all the collision jobs. I should maybe use an enabled component to flag rotations for the spawn point movement system.

But the bigger culprit here is the transform hierarchy update. I haven't implemented the equivalent of Extreme Transforms yet, and so this is demanding as much frame time as the collision jobs. I need to decide how I want to fix that though because that technique requires chunk component structural changes unless I just query the depths every frame.

The other potential big optimization will be converting spawn point clearance to use layer queries with better early-outs. It'll be interesting if Burst can early-out or if I have to do it manually.