Update to 2022.2.12 and Entities 1.0-Pre.47
Warning: The subscene serialization bug is present in this update. If you encounter the issue and play mode does not work with the subscene closed, open the subscene before entering play mode.

This update was necessary because the newer versions support new features that are being developed for the Latios Framework to better enhance overall stability.