Fix vertex skinning, a tiny baking bug, and start Myri stacks
Lots of changes in this update, with the biggest being the WIP Myri Stacks feature.

There's a small bugfix with SmartBlobberBakingGroup where now people are getting systems to register late instead of early. I have written and removed this code so many times experimenting with the upgrade to Pre.47 that I was bound to end up with a borked up version. At least it is fixed now.

Then there's the conversion of vertex skinning to use ByteAddressBuffer instead of StructuredBuffer. Now it is working again. I did discover what seems to be some sort of change filter bug with transforms when live baking. That needs investigation.

Lastly, there's all the stuff with Myri. But it is still in progress, so I won't talk about it now. What I will say is that next up I probably need to roll out a custom allocator and allocator helper pair that can be constructed off thread.