Burst more things (plus lots of stability updates)
We're getting much closer to a releasable state. On top of the previous alpha updates, we now have Burst-compiled Fluent Queries and Burst-schedulable FindPairs. That didn't have much of an impact on performance, but it is still nice to have.

Transforms baking has been reworked to add an additional user entry-point as well as cleaning up WorldTransform sync issues via a brute-force update.

And lastly, scripting defines have been added for Transforms and decorated throughout the framework so that now it is much more intuitive how to add support for other transform modes like the Unity compatibility mode.

One really weird thing is that motion history seems to be updating at the wrong time in the frame. It seems like OrderLast is being treated like OrderFirst. We'll see if that gets fixed next Entities update, or else that will need a workaround.

Lots still to do.