Improvements and Cleanup
Unfortunately, FindObjects optimization for spawn points didn't work out. But the code is left there just in case anyone wants to play with it. Otherwise, this mostly contains a couple of CollisionLayer renames for consistency, and some other minor cleanups and improvements as the framework is getting ready for beta.

There are three outstanding bugs that will hopefully be fixed in the next ECS release. There is a bug with subscene serialization and blob assets and/or shared components. There is a bug with baking optimized skeletons (this is a Unity Editor bug that has been fixed in later versions than the ones I have been developing with). And there is a system update sort order bug which results in MotionHistoryUpdateSystem running earlier than it is supposed to.