Entities 1.0.8 Update
Entities is no longer pre-release, but apparently that doesn't make it bug-free. It is still quite leaky on memory. Fortunately, it seems most of the subscene issues are gone, though I haven't tested it very much.

The big change was with bootstraps and SystemTypeIndex. I'm not a fan of the way BootstrapTools has evolved to cope with all of Unity's evolutions, but the rewrite will probably have to wait until 0.8.

Also, I moved all the bakers to use TransformUsageFlags. I accidentally wrote a bug in one of the LSSS bakers which lead to discovering a bug in Transform Baking with the flags that I got to fix.

And lastly, I got to stomp out a bunch of memory leaks. They were all once-per-world type of leaks, and all of small allocations. So they didn't really affect builds at all. But it is nice not having them in the Editor. One of the big wins is that one of these leaks helped me find the DSP leaks that always got warned about. Myri no longer has warnings when exiting play mode! Finally!

Next up, I have to debug Kinemation, especially optimized skeletons. That will either not take very long, or completely blow up in my face. We'll see.