Update to 0.7.2 plus command buffer refactors
Command buffers were a problem with their inconsistent behavior with and without Burst which was too volatile to memory leaks and other issues. Instead of relying on jobs for Burst, I'm now relying on static methods that can be Burst-compiled. It is much cleaner and more predictable. And no more leaks now that Allocator.Temp is the goto allocator.

This change will likely land in the 0.7.3 along with an empty folder getting removed.