New FindPairs tools from sphere simulation sample
The sphere simulation sample made a few issues apparent that were pretty trivial to fix. The first was that there wasn't a RefRW access for PhysicsComponentLookup. Using RefRW is preferred because it removes overhead of read-modify-write operations. The second change was adding a Physics lookup for TransformAspect, since that is a really common operation. I may need to consider adding lookups for the Kinemation aspects too, but I'll wait until those features are requested. I'm hoping that there will be an easier way to do this in the feature with generics. Or maybe I write a custom source generator for it?

There's also the new Substep API. The idea is that you call the Substep method as the "collection" in a foreach statement, and that will run your loop multiple times with substepped time values that will all add up to exactly deltaTime.