Add feature codes for Psyshock DistanceBetween
Feature codes will allow for contact generation algorithms to quickly load up context for finding contacts given a ColliderDistanceResult and the inputs that generated it.

I'm still not sure if I'm on the right track with this, but the code is there now.

I'm probably going to let this new changes settle before I investigate contact generation. Not sure what I will do in the meantime.