Fix sceneBlackboardEntity execution order and add GameObjectEntity authoring
I had to completely rethink the execution order for sceneBlackboardEntity and OnNewScene callbacks, and ended up lazy-initializing things in the merging system. This does mean that the merging system updates every frame now. Hopefully that doesn't cause any major performance problems in the future. Otherwise, I might gate it with a ShouldUpdateSystem to solve it.

As for GameObjectEntity, I think I got the authoring pieces done. Next up is the systems, and then wiring up LSSS to use the new functionality.