GameObjectEntity is working!
CameraManagerCameraSyncHackSystem is dead! The new paradigm is in place and is functioning as it should. Unfortunately, I couldn't quite allow for host prefabs, as the reference implementation caused new GUIDs to be generated whenever the subscene closed. I think I'm close though, so I will have to play around with it more in the future.

Next is Mecanim events. And then beta!