Begin 0.9 development
This started out by updating to Entities 1.1.0-exp.1 which caused subscenes to break. The editor version was updated in the process of resolving the subscenes issue.

Second, some fixing needed to happen for Entities Graphics, as entities were receiving the wrong materials. This exploration resulted in the discovery of IN-57221 which remains unresolved. In the meantime, LSSS is working again, but other projects may not be so fortunate.

There's still a lot of cleanup left to do, especially with Fluent Queries, but also with Kinemation flags and culling parameters.