Add CPU Skinning and Normalization
If the Latios Framework is going to be used for high-fidelity character animation, than it is crucial it provides a low-level deformation pipeline for those who need it to do advanced visual effects. For now, this pipeline is provided CPU-side with Burst. In the future, a GPU pipeline will also be provided.

Exposing the pieces for this pipeline was complicated. I needed to expose parts of collection components as well as parts of internal component types. My solution for this was IAspects and a new ICollectionAspect interface.

Also, there are so many edge cases with duplicate vertices for mesh normalization purposes. Coming up with a compact-ish data structure and plugging all the gaps was extremely challenging. Hopefully this API is somewhat useful for such issues.