Add thread indices test
Turns out that JobsUtility.ThreadIndex takes about the same amount of time as an Rng random int. The job that queries from JobsUtility takes twice as long as the job that uses injected index. For the unfamiliar, that might sound scary, but this is actually quite good. It means that the mechanism for querying the index likely doesn't require hitting the kernel to compare native thread IDs. Unity has a fast implementation and the overhead primarily comes from the function call itself. This is totally usable for batch allocations.

Next up will probably be the thread allocators and job allocators, and then I will either investigate determinism or FindPairs optimizations, depending on the survey results progress.