Bugfixes and ThreadStackAllocator
There were a few bugs discovered regarding inertial blending and building for a NetCode project. Those bugs have been squashed now.

In addition, I have the code written for the new ThreadStackAllocator, which should provide a much-needed utility for temporary allocations in that methods. It is very unsafe though, so it is mostly suited for framework code or other libraries written by people who know what they are doing.