Add Cross Caching for ParallelByA and fix scheduling
Another massive improvement! By bringing down the job count, scheduling improved significantly. Unfortunately, radars are still pretty expensive on the cross-bucket, so we had to keep them divided by faction. But that in a weird way opened up other optimizations.

Cross-caching was added to ParallelByA, though it didn't actual help, as the processor itself was expensive. But in the process I have added UnsafeIndexedBlockList and refactored UnsafeParallelBlockList to use it.

LSSS has a few bottlenecks now. One of them is the radar logic. One of them is bandwidth for all the transforms, especially motion history for motion vectors. And Unity's engine code handling the reception of render batches is very time-consuming on the main thread.