Fixed Myri bugs and start PairStream
There were two bugs with Myri. First, it was playing everything at half volume. And second, it was doubly-applying channel weights. This was fixed by removing a bunch of logic in the culling and weighting that didn't need to be there.

Also, I did a little bit of consistency cleanup in UnitySimMotion.cs and I've started the PairStream implementation, though I haven't gotten ver far yet.