Implement PairStream
I needed to make modifications to UnsafeIndexedBlockList to support concatenation and ref reading. As it turns out, AllocatorManager ensures greater than byte alignment, so reading by ref works as long as the element size has the element's desired alignment as a factor.

I also had to rework FindPairsResult and FindPairsBucketContext to contain the bucket indices since those need to be remapped to stream indices.

Next up is the Physics.ProcessPairs implementation. There's plenty to do there as well.