Change PairAllocator to Pair for full pair read-write
I decided to rework PairAllocator to instead be the Pair type that is passed to the user for full access. I managed to discover and fix several safety issues in the process. Now I just need to work out the names for the job schedulers and interfaces. But maybe I'll do the enumerator first?