Add ForEachPair mechanisms and clean up PairStream
ForEachPair became the name of choice from the community. And now it is all written. The islanding algorithm was a bit tricky, but hopefully this implementation is fast and accurate.

I've also documented PairStream and all associated APIs, and found and fixed a couple of bugs in the process.

While updating the Early Initers for scheduling from Burst, I realized I had added the preservation to the wrong types, which meant preservation wouldn't work correctly. I fixed that now.

Next up is to put this all together in Free Parking!