Update to Unity 6 Preview and fix compilation and GC issues
In Unity 6 Preview, SortedSet.Add and StringBuilder.Append(int/float) allocate GC now. Plus, there were a few compilation warnings and other tiny issues that needed to be resolved.

Unfortunately, builds are a little unstable right now. I don't know yet if this is a Burst bug or an Entities bug. But I got a crash when adding a tag component to the UniversalQuery. And Burst is not running in development builds. So I will probably wait on publishing new LSSS versions.