Fix PairStream indexing
I completely overlooked the case where there are a bunch of threads trying to all write to one of the three cross-bucket indices when only the cross-bucket is writable. Well now there's a lot more indices to write to, and I have reworked how indexing is handled to all be gathered into a single file. Hopefully this makes everything more stable.

There's also a tiny fix for Kinemation where multiple samples with weights summing to 1f resulted in quaternion normalization being skipped. So I removed the fast path and now normalization always happens.

Time for a patch, and then on to optimizations and whatever else.