Add Unika resolvers and start structural changes
Unika development continues. The resolvers are now done, including the cached wrapper. The runtime add allocation routine is also written, leaving just removal, the ScriptCommandBuffer, and a whole bunch of quality-of-life APIs. And then there's the source generators. No idea which of these I will tackle next. But there's still plenty to do!