Add compute shader unmanaged APIs
I decided to start with a direct API for compute shaders rather than try to do CommandBuffers. If it turns out to be a big performance hit, I can add the latter later. But the immediate API is much more easily debugged.

Next up is to start migrating GraphicsBufferBroker to this new API.