Add bugfixes and improvements to NetCode QVVS
A templates bug was fixed, as was an issue that resulted in the server world being made default instead of the client. Now instead of exclusively relying on a first-come-first-serve, the client world can specify itself as default. This fixed GameObjectEntity in NetCode.

Also, scene management was enabled for NetCode, and I've added InputDeltaUtilities based on the Online FPS character controller sample. I have no idea why that isn't part of an official package, so I put it in Core.