Optimized skeleton cache plus bugfixes
Two bugs got fixed. The first was culling not working correctly due to WriteGroup issues. The second was dispatch systems not showing up in the Systems window in Unity 6.

As for the optimized skeleton cache, the initial implementation is very crude, and I still need to think up a better API for it. But it is at least a start.