More Unique Meshes plus blend shape safety
There appears to be a situation where Unity fails to acquire the blend shape buffer for meshes that have blend shapes. We can detect this and fail a little more gracefully.

As for unique meshes, I now have the MeshData generation complete, minus normal and tangent recalculation. However, this job is not very optimized, nor is the resulting mesh. So these may be future investigations.