Cleanup after the jam
A few things happened during the jam that were worth cleaning up. First, people were confusing framework components with built-in components. So I changed how they show up to be a little more obvious.

Second, I reproduced the blob asset leak. Turns out, this leak has been around ever since the internalization of BlobAssetComputationContext. Why it gets warned so infrequently is still a mystery that I guess will never be solved. I solved the leak instead.

And lastly, there was a change to blend-shapes to partake in the never-ending game of whack-a-mole of complex entities being spawned at weird times.