Allow culled LOD1 in append and support 3-LOD append
MMI Range LODs are very performant, but being limited to 2 levels that are always visible was not enough. So time to allow for 3 levels, and allow the lowest level to also be a culled level so that entities can fade out at a certain distance. We'll see how much extra mileage this gets us.