Retrieve system fonts from Unity
Turns out, I can retrieve the system font list from within a job. However, all the data is managed, so I had to convert it all into unmanaged data and send it to a Burst static method. Nothing is implemented beyond that point.

Also, there was a request to add direction helpers to TransformQvvs.