Continue making a mess in Calligraphics
There's so much going on in Calligraphics that is messy. I'm just scoping out what all data is needed by faking out the inputs to glyph generation. Next, I need to get the tables and shape stream data sorted out so that I can clean up glyph generation and move on to shaping.