Revert Calligraphics, add LifeFX subgraphs and Myri pitching
Calligraphics is now back to normal again, as the new Harfbuzz revision is no longer planned for 0.13, even though development continues on the side.

I've added some VFX Subgraphs for LifeFX in sample-form, although some of them break Unity's code generator for mysterious reasons. I also ended up fixing some ShaderLibrary bugs in the process.

And Myri now got AudioSourceSampleRateMultiplier, which applies a multiplier to the sample rate to alter timing and pitch. It was a requested feature.