Switch Brickwall Limiter to new parameterized algorithm
It is time for a new parameterized and customizable Brickwall Limiter algorithm. I modified the effect stack version to operate per sample rather than per audio frame, as this allows varying the pregain and volume on a per-sample basis. Also, I reconfigured the dB limit to be a volume control, which simplifies both the code and usability.