Inherit and improve more DrawCommandGeneration
I decided that I wanted to change how LOD Crossfade pointers work, so I decided to inherit even more of the stack. But with this effort, I was able to identify some bugs and other issues in Unity's code and fix them.

For example, some of the per-thread structures weren't aligned to cache lines, but instead used a little more than 1, which meant either dual cacheline loads and in one case this caused false sharing. I also replaced a lot of TempJob allocations with RewindableAllocator allocations, and managed to completely eliminate the need for all disposal jobs in the pipeline.

Next up, I want to change up ChunkVisibility and remove TransformStream, and ultimately remove the LOD Crossfade pointer hashmap by putting it in the ChunkVisibility stream directly.