Add TrueSim to Psyshock and apply 0.13.2 fixes
There were a number of bugfixes that made it into 0.13.1 and 0.13.2. But aside from that, it is time to clean up some Psyshock APIs. Ballistics got moved out of the Physics class and into a new TrueSim class along with some other well-established formulas for center of mass and inertia tensors.