Optimize EmitDrawCommandsJob and add Unity 6.0 compatibility
It turns out, there wasn't many code points that were broken when I accidentally imported the framework into a 6.1 project. So I've wrapped all those points in conditionals and now have backwards compatibility.

As for the optimization, I'm writing an Optimization Adventure on that. This is simply the first step, which resolves a Mesh LOD performance regression. Still more work to do!