Fixed Myri falloff issues
Myri had an intrinsic issue where changes in volume from the falloff calculations would cause an instantaneous change in magnitude between samples. And this resulted in an impulse spike that caused nasty clicking sounds. The solution is to capture a couple extra samples each frame, and use those to compare against the next update's version. If they don't match, the samples are offset by the difference faded to 0 over 128 samples. This makes the waveform continuous.

Also, falloff disabled mode was broken due to a stale EntityQuery.