Optimize LSSS on new computer
The new computer has a lot more threads, and consequently, the performance dynamics are a bit different.

Turns out, radar scans were not being distributed efficiently, and it is cheaper to just perform FindObjects queries for each one in-place. Also, there were a couple of jobs where we could cull chunks.

Bandwidth is a much bigger problem than FindPairs when more cores and threads are involved, as are structural changes. There might be enough for another optimization adventure for those, but we'll see if I feel like it.