Collection Component API and more graphics unmanaging
There was a feature request to add a readOnly version to the Collection Component API that returns an explicit JobHandle. And I continue with making more things in the Unmanaged struct of LatiosEntitiesGraphicsSystem actually unmanaged. Next up, I need to rework the culling and dispatch callbacks to remove the system references.